add state management
The top-level ansi package is responsible for consuming the actions
generated by the parser and modifying some internal state. Any
interesting work related to printing is delegated to a `Printer`
interface (which can Print bytes with certain styles applied, as well as
Clear bytes to the right of a position on a line).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>