rename Buffered to InMemory
Rather than stuffing all of the logic for basic in-memory printing and
communicating with Storage into a single type, split this
responsibility. Buffered will be implemented using InMemory for dealing
with the segment of lines that are currently in memory.

This is also convenient since I can definitely see use cases where it'd
be nice to just store all the lines in memory - my plan for this
originally was to expose some sort of dummy "MemoryStorage" and to use
the Buffered output.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>