disallow moving cursor beyond "screen"
As we don't really have the concept of a screen, keep track of the
maximum size seen so far (with a default of 48L x 80C). The cursor
cannot be moved beyond this. The width of the screen can be expanded by
printing beyond the 80C mark, and the height by linebreaks.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>