make ansi.New wire things up
It's really awkward and counter-intuitive for ansi.New to return a
State, and require the user to wire up the parser and output. This
change makes ansi.New return an ansi.Log, which has a parser and state
baked in. The output must still be wired separately, since not all
implementations will expose Lines, but this is often desirable.

Also, introduce some functional options.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>